refactor routing code into rolo.routing module
