+++ tests/test_pydantic.py
@@ -2,8 +2,9 @@ import pydantic
 import pytest
 from werkzeug.exceptions import BadRequest
 
-from rolo import Request, Router, dispatcher, resource
-from rolo.dispatcher import handler_dispatcher
+from rolo import Request, Router, resource
+from rolo.routing import handler as routing_handler
+from rolo.routing import handler_dispatcher
 
 
 class MyItem(pydantic.BaseModel):
@@ -153,7 +154,7 @@ class TestPydanticHandlerDispatcher:
         assert router.dispatch(request).get_json(force=True) == {"item": None}
 
     def test_with_pydantic_disabled(self, monkeypatch):
-        monkeypatch.setattr(dispatcher, "ENABLE_PYDANTIC", False)
+        monkeypatch.setattr(routing_handler, "ENABLE_PYDANTIC", False)
         router = Router(dispatcher=handler_dispatcher())
 
         def handler(_request: Request, item: MyItem) -> dict:
